% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNet.R
\name{addReaction}
\alias{addReaction}
\title{For adding reactions to a pre-existing network}
\usage{
addReaction(
  network,
  reactants,
  products,
  reversible = FALSE,
  reactionName = NULL
)
}
\arguments{
\item{network}{xml_document. An SBML file which defines the chemical reaction
network. It should have been originally created with \code{makeNetwork}.}

\item{reactants}{Named Integer vector. Each element specifies the
stoichiometric coefficient of a reactant in the reaction, and the name of the
element specifies the name of the reactant species.}

\item{products}{Named Integer vector. Each element specifies the
stoichiometric coefficient of a product in the reaction, and the name of the
element specifies the name of the product species.}

\item{reversible}{(optional) Logical. Default FALSE. Whether or not the input
reaction is reversible.}

\item{reactionName}{(optional) Character. Default NULL. The ID of the
reaction for the SBML file. If not provided, the SBML ID for the reaction
will be "Rn+1", where n is the number of reactions already in the network
before this was one added.}
}
\value{
xml_document. The input network modified to contain the input
reaction.
}
\description{
This function adds reactions to a pre-existing chemical reaction network as
defined by the \code{makeNetwork} function. The reactions are added in a way
that keeps the model xml object completely compatible with System Biology
Markup Language (SBML) Level 3, Version 2. Kinetic laws are not included in
SBML file as in crnRACIPE, all reactions are assumed to follow mass-action
kinetics.
}
\section{Related Functions}{


\code{\link{makeNetwork}}
}

\examples{
example <- makeNetwork("example")
example <- addReaction(network = example, reactants = c(A = 1, B = 1),
                       products = c(AB = 1), reversible = TRUE,
                       reactionName = "Fusion")
rm(example)
}
