% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{annotation,cRacipeSE-method}
\alias{annotation,cRacipeSE-method}
\title{A method to get the annotation}
\usage{
\S4method{annotation}{cRacipeSE}(object, ...)
}
\arguments{
\item{object}{cRacipeSE object}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
character
}
\description{
A method to get the annotation
}
\examples{

cSet <- cRacipeSE()
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cracipeNetwork(cSet) <- wilhelm
ann <- annotation(cSet)
rm(cSet, wilhelm,ann)
}
