% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{configData}
\alias{configData}
\title{Configuration Data}
\format{
List of vectors
}
\usage{
configData
}
\description{
It contains vectors of parameters defining the simulation. The
simParams vector contains the number of models to run (numModels). The
hyperParams vector contains the min and max values for kinetic rate sampling
(kMin, kMax) and the min and max values for initial condition sampling
(icMin, icMax). The options vector contains logical parameters such as
whether or not to integrate the initial conditions (integrate).
}
\keyword{datasets}
