% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeConfig}
\alias{cracipeConfig}
\alias{cracipeConfig,cRacipeSE-method}
\title{A method to access the simulation hyperparameters}
\usage{
cracipeConfig(.object)

\S4method{cracipeConfig}{cRacipeSE}(.object)
}
\arguments{
\item{.object}{cRacipeSE object}
}
\value{
list
}
\description{
The hyperparameters like number of models, range from which
parameters are to be sampled, simulation time etc.
}
\examples{
cSet <- cRacipeSE()
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cracipeNetwork(cSet) <- wilhelm
cracipeConfig(cSet)
rm(cSet)
}
