% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeConverge}
\alias{cracipeConverge}
\alias{cracipeConverge,cRacipeSE-method}
\title{A method to get the convergence results}
\usage{
cracipeConverge(.object)

\S4method{cracipeConverge}{cRacipeSE}(.object)
}
\arguments{
\item{.object}{cRacipeSE object}
}
\value{
DataFrame
}
\description{
Gathers the convergence and speed of convergence for each model
and initial condition.
}
\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cSet <- racipeCRN::cracipeSimulate(network = wilhelm, numModels = 100, numIC = 3)
cd <- cracipeConverge(cSet)
rm(cSet, cd)
}
