% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeConvergeDist}
\alias{cracipeConvergeDist}
\alias{cracipeConvergeDist,cRacipeSE-method}
\title{A method to visualize convergence distributions}
\usage{
cracipeConvergeDist(.object, plotToFile = FALSE)

\S4method{cracipeConvergeDist}{cRacipeSE}(.object, plotToFile = FALSE)
}
\arguments{
\item{.object}{cRacipeSE object generated by \code{\link{cracipeSimulate}}
function.}

\item{plotToFile}{(optional) logical. Default \code{FALSE}. Whether to save
plots to a file.}
}
\value{
\code{cRacipeSE} object
}
\description{
Generates a time plot of the ratio of converged models as the simulation went
on. These visualizations can be used for evaluating the ideal simulation run
time, as defined by the number of iterations, number of steps per iteration,
and the step size. The final proportion of converged concentrations at the
end of the simulation is also computed and placed in the output metadata. If
the final proportion is greater than 99\%, the earliest iteration with 99\%
convergence is also reported to the metadata.
}
\section{Related Functions}{


\code{\link{cracipeSimulate}},  \code{\link{cracipePlotData}}
}

\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
\dontrun{
cSet <- racipeCRN::cracipeSimulate(network = wilhelm, numModels = 20,
integrateStepSize = 0.001, numIterations = 30)
cSet <- cracipeConvergeDist(cSet)
}
}
