% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cracipeGenParamNames}
\alias{cracipeGenParamNames}
\title{Generate parameter names for a network}
\usage{
cracipeGenParamNames(network)
}
\arguments{
\item{network}{cRacipeSE object}
}
\value{
character vector
}
\description{
Generate parameter names for a network
}
\examples{
cSet <- cRacipeSE()
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cracipeNetwork(cSet) <- wilhelm
paramNames <- cracipeGenParamNames(cSet)

}
