% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeIC<-}
\alias{cracipeIC<-}
\alias{cracipeIC<-,cRacipeSE-method}
\alias{cracipeIC-set}
\title{A method to set the initial conditions}
\usage{
cracipeIC(.object) <- value

\S4method{cracipeIC}{cRacipeSE}(.object) <- value
}
\arguments{
\item{.object}{cRacipeSE object}

\item{value}{DataFrame containing the initial conditions, dimensions should
be (# species) rows by (numModels*numIC) columns}
}
\value{
A cRacipeSE object
}
\description{
Set the initial conditions
}
\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cSet <- racipeCRN::cracipeSimulate(network = wilhelm, numModels = 10,
integrate=FALSE)
ics <- cracipeIC(cSet)
cracipeIC(cSet) <- ics
rm(cSet, ics)
}
