% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeIC}
\alias{cracipeIC}
\alias{cracipeIC,cRacipeSE-method}
\title{A method to get the initial conditions used for simulations}
\usage{
cracipeIC(.object)

\S4method{cracipeIC}{cRacipeSE}(.object)
}
\arguments{
\item{.object}{cRacipeSE object}
}
\value{
DataFrame with numModels columns and numSpecies rows.
}
\description{
The initial conditions of each of the models.
}
\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cSet <- cracipeSimulate(wilhelm, integrate = FALSE, numModels=20)
ic <- cracipeParams(cSet)
rm(ic,cSet)
}
