% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeNetwork<-}
\alias{cracipeNetwork<-}
\alias{cracipeNetwork<-,cRacipeSE-method}
\alias{cracipeNetwork-set}
\title{Initialize the Network}
\usage{
cracipeNetwork(.object) <- value

\S4method{cracipeNetwork}{cRacipeSE}(.object) <- value
}
\arguments{
\item{.object}{cRacipeSE object.}

\item{value}{xml_document or character. SBML file or .tpo text file containing
the network topology. If it is a character, it should be a path to the desired
txt file. See README for more information about the format of .tpo files.}
}
\value{
cRacipeSE
}
\description{
Initialize a cRacipeSE object using an SBML file containing a chemical
reaction network (CRN). All reactions are assumed to have mass-action
kinetics, and any defined rate law in the SBML file will be ignored. Please
see \link{makeNetwork} and \link{addReaction} for methods to make a
compatible SBML file
}
\examples{
cSet <- cRacipeSE()
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cracipeNetwork(cSet) <- wilhelm
rm(cSet, wilhelm)
}
