% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeParams<-}
\alias{cracipeParams<-}
\alias{cracipeParams<-,cRacipeSE-method}
\alias{cracipeParams-set}
\title{A method to set the ensemble kinetic parameters}
\usage{
cracipeParams(.object) <- value

\S4method{cracipeParams}{cRacipeSE}(.object) <- value
}
\arguments{
\item{.object}{cRacipeSE object}

\item{value}{DataFrame containing the parameters. Dimensions should be
(numModels) rows by (numReactions) columns.}
}
\value{
A cRacipeSE object
}
\description{
Set the parameters
}
\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cSet <- racipeCRN::cracipeSimulate(network = wilhelm, numModels = 20,
integrate = FALSE)
parameters <- cracipeParams(cSet)
cracipeParams(cSet) <- parameters
rm(parameters, cSet)
}
