% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeParams}
\alias{cracipeParams}
\alias{cracipeParams,cRacipeSE-method}
\title{A method to access the kinetic rate constants for each model in the
ensemble}
\usage{
cracipeParams(.object)

\S4method{cracipeParams}{cRacipeSE}(.object)
}
\arguments{
\item{.object}{cRacipeSE object}
}
\value{
A data.frame with numReactions columns and numModels rows.
}
\description{
Gets the kinetic rate constants for each model in the form of
a data.frame.
}
\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
cSet <- cracipeSimulate(wilhelm, integrate = FALSE, numModels=10)
parameters <- cracipeParams(cSet)
rm(parameters,cSet)
}
