% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{cRacipeSE-class}
\alias{cRacipeSE-class}
\alias{.cRacipeSE}
\title{cRacipeSE}
\description{
An S4 class for Random Circuit Perturbation (RACIPE)
simulations of chemical reaction networks.
Extends the \link{SummarizedExperiment} class.
}
