% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{cRacipeSE}
\alias{cRacipeSE}
\title{cRacipeSE constructor}
\usage{
cRacipeSE(
  .object = NULL,
  assays = SimpleList(),
  rowData = NULL,
  colData = DataFrame(),
  metadata = list(),
  ...
)
}
\arguments{
\item{.object}{(optional) Another cRacipeSE object.}

\item{assays}{(optional) assay object for initialization}

\item{rowData}{(optional) rowData for initialization}

\item{colData}{(optional) colData for initialization}

\item{metadata}{(optional) metadata for initialization}

\item{...}{Arguments passed to SummarizedExperiment}
}
\value{
cRacipeSE object
}
\description{
Create an RacipeSE object. cRacipeSE is an S4 class for
Random Circuit Perturbation (RACIPE) simulations of chemical reaction
networks in which a large number of models with randomized parameters are
used for simulation of the circuit. Each model can be considered as a sample.
It extends the \link{SummarizedExperiment} class to store and access
the network, simulated expressions, parameters, intial conditions and
other meta information.
SummarizedExperiment slot assays is used for storing simulated
species expressions. The rows of these
matrix-like elements correspond to various species in the circuit and columns
correspond to models.
}
\examples{
cSet <- cRacipeSE()

}
