% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cracipeSimulate.R
\name{cracipeSimulate}
\alias{cracipeSimulate}
\title{Simulate a chemical reaction network}
\usage{
cracipeSimulate(
  network,
  config = config,
  numModels = 2000,
  numIC = 1,
  outputPrecision = 12,
  kMin = 0,
  kMax = 10,
  icMin = 0,
  icMax = 100,
  stepper = "E",
  integrateStepSize = 0.02,
  convergThresh = 1e-12,
  numStepsIteration = 500,
  numIterations = 25,
  integrate = TRUE,
  genParams = TRUE,
  genIC = TRUE,
  plots = FALSE,
  plotToFile = FALSE,
  normalized = FALSE
)
}
\arguments{
\item{network}{xml_document or cRacipeSE object or character containing the
file path to a text file with the network's topology. xml_document objects
must be in SBML format.}

\item{config}{(optional) List. Contains the simulation parameters and
hyperparameters. It should contain the vectors simParams, hyperParams, and
options.}

\item{numModels}{(optional) Integer. Default 2000. Number of random models
to be simulated.}

\item{numIC}{(optional) Integer. Default 1. Number of initial conditions to
simulate for each model.}

\item{outputPrecision}{(optional) integer. Default 12.
The decimal point precision of the output parameters and concentrations.}

\item{kMin}{(optional) numeric. Default 0. Minimum of uniform distribution
for sampling kinetic rate constants.}

\item{kMax}{(optional) numeric. Default 10. Maximum of uniform distribution
for sampling kinetic rate constants.}

\item{icMin}{(optional) numeric. Default 0. Minimum of uniform distribution
for sampling initial conditions.}

\item{icMax}{(optional) numeric. Default 100. Maximum of uniform distribution
for sampling initial conditions.}

\item{stepper}{(optional) character. Default "E". Defines the integration
method used for simulation. Use "E" for the 1st-Order explicit Euler method.}

\item{integrateStepSize}{(optional) numeric. Default \code{0.001}. step size for
integration.}

\item{convergThresh}{(optional) numeric. Default \code{1e-12}. The threshold
for simulation convergence.}

\item{numStepsIteration}{(optional) integer Default \code{500}. The number of
integration steps between per iteration.}

\item{numIterations}{(optional) integer. Default \code{150}. The total
number of convergence test iterations to run per model.
initial condition in deterministic simulations.}

\item{integrate}{(optional) logical. Default \code{TRUE}. Whether to
integrate the differential equations or not. If \code{FALSE}, the function
will only generate the parameters and initial conditions. This can be used
for setting the kinetic rate constants and initial conditions and then
adjusting them as needed.}

\item{genParams}{(optional) logical. Default \code{TRUE}. Whether or not to
generate kinetic parameters for each model. If set to \code{FALSE}, the
network should be a cRacipeSE object with preexisting parameters to use.}

\item{genIC}{(optional) logical. Default \code{TRUE}. Whether or not to
generate initial conditions for each model. If set to \code{FALSE}, the
network should be a cRacipeSE object with preexisting initial conditions to
use.}

\item{plots}{(optional) logical. Default \code{FALSE}. Whether or not to
plot the simulated data using \code{\link{cracipePlotData}} and
\code{\link{cracipeConvergeDist}}.}

\item{plotToFile}{(optional) logical. Default \code{FALSE}. Whether or not
to save the plots to a file.}

\item{normalized}{(optional) logical. Default \code{FALSE}. Whether or not
to noemalized the concentration values.}
}
\value{
cRacipeSE object
}
\description{
Simulate an ensemble of models for a chemical reaction network
using its topology as the only explicit input. The ensemble is randomly
generated.
}
\section{Related Functions}{


\code{\link{cracipeNetwork}}
}

\examples{
wilhelm <- system.file("extdata", "wilhelm.tpo", package = "racipeCRN")
cSet <- cracipeSimulate(wilhelm)

}
