% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cracipeUniqueStates}
\alias{cracipeUniqueStates}
\alias{cracipeUniqueStates,cRacipeSE-method}
\title{A method for grabbing unique states}
\usage{
cracipeUniqueStates(.object, roundDigits = 3)

\S4method{cracipeUniqueStates}{cRacipeSE}(.object, roundDigits = 3)
}
\arguments{
\item{.object}{cRacipeSE object generated by \code{\link{cracipeSimulate}}
function.}

\item{roundDigits}{(optional) integer. Default \code{3}. Number of digits for
concentration rounding.}
}
\value{
\code{list} object. Element i of the list is a data frame containing
the unique expressions of model i in the input cRacipeSE object
}
\description{
This method selects the unique expression states for every model
in a racipeCRN object. Non-converged expressions are  filtered out using the
simulation convergence data. To minimize error from numerical integration,
concentration values are rounded before comparison.
}
\section{Related Functions}{


\code{\link{cracipeSimulate}}
}

\examples{
wilhelm <- xml2::read_xml(system.file("extdata", "wilhelm.xml", package = "racipeCRN"))
\dontrun{
cSet <- racipeCRN::cracipeSimulate(network = wilhelm, numModels = 20,
integrateStepSize = 0.001, numIterations = 30)
stateList <- cracipeUniqueStates(cSet)
combinedStates <- do.call(cbind, stateList)
}
}
