% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNet.R
\name{makeNetwork}
\alias{makeNetwork}
\title{A method to quickly build an SBML frame}
\usage{
makeNetwork(netName)
}
\arguments{
\item{netName}{character. Name of the newly made chemical reaction network}
}
\value{
xml_document
}
\description{
A method to quickly make an SBML level 3, version 2, List object using xml2.
It builds the model and adds a singular compartment for the purpose of adding
species. It uses the minimum required parameters and it is designed solely for
building chemical reaction networks with mass action kinetics in a single
compartment.
}
\section{Related Functions}{


\code{\link{addReaction}}
}

\examples{
crn <- makeNetwork("example")
rm(crn)
}
