% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racipeCRN.R
\docType{package}
\name{racipeCRN}
\alias{racipeCRN-package}
\alias{racipeCRN}
\title{racipCRN: A package for the random circuit perturbation of chemical reaction
networks.}
\description{
racipeCRN is a tool for analyzing chemical reaction networks (CRNs) under
changing kinetic parameters. With just the CRN topology, racipeCRN
generates an ensemble of models for the CRN using uniform random sampling
of the rate constant parameters.
}
\author{
\strong{Maintainer}: Aidan Tillman \email{aidan.tillman18@gmail.com}

}
